Moved new noise reduction to audio_nr.c and some small changes - Moved audio noise reduction into own file - renamed alternate_NR to USE_ALTERNATE_NR - slighty optimization on spectrum scaling code, nothing serious
