+++ mchf-eclipse/hardware/uhsdr_board.h
@@ -17,10 +17,10 @@
 // some special switches
 //#define 	DEBUG_BUILD
 
-#define alternate_NR
+#define USE_ALTERNATE_NR
 //#define debug_alternate_NR
 
-#ifdef alternate_NR
+#ifdef USE_ALTERNATE_NR
 #define NR_FFT_SIZE 128
 #endif
